Various tests and fixes
Operand typeinfos now use their internal names as unique identifiers, allowing you to compare two typeinfos via `a == b`. This is because in MSVC, `&a == &b` is flaky given `constexpr` instances, when comparing references or pointers generated from different translation units.

All typeinfos were changed from `static constexpr` to `inline static constexpr` in a vain attempt at getting MSVC to behave sensibly, before the above change was eventually settled for.

The `game_variant` class now actually attempts to read and store `megalo_variant_data`. (Whoops!)

Fixed the string table's loading of 0 and -1 string offsets.

Documented bitnumber's handling of signed types. That said, this functionality is awful and error-prone, especially when dealing with enums that have no explicit underlying type (as well as structs with `underlying_type` `int`). We should have the presence or absence of a sign bit be explicitly indicated in bitnumber params instead.

Several enums were given unsigned underlying types, as a workaround for bitnumber's poorly-conceived sign bit handling.

Created a natvis file for `halo::util::fixed_string`.

Fixed `halo::util::indexed_list`'s iterator creation code, so that creating `end()` doesn't trigger MSVC debug assertions (re: access past the end of an array).

+++ native/src/ReachVariantTool/halo/reach/trait_set.cpp
@@ -23,7 +23,7 @@ namespace halo::reach {
             if constexpr (bitreader::has_load_process) {
                auto message_data = halo::common::load_process_messages::player_trait_out_of_bounds_message_content{
                   .trait       = Id,
-                  .value       = vi,
+                  .value       = (int)vi,
                   .allowed_min = info.min,
                   .allowed_max = info.max,
                };
